requirements.txt
flask
py-cord==2.4.1
websockets
pandas
numpy
ta
python-dotenv
